fgerert

+++ a60-四/src/main/scala/四法/Counter7.scala
@@ -68,7 +68,7 @@ package 测试1 {
       val yuan        = (i - 1).toDouble / i
       var num: Double = 0
       for (num1 <- 1 to 4000) {
-        num += math.pow(yuan, num1) / num1.toDouble
+        num += math.pow(yuan, num1) / num1
       }
       assert(math.abs(math.log(i) - num) < 0.000001d)
     }
